+++ test/test_dumper.py
@@ -2019,7 +2019,8 @@ _ihm_2dem_class_average_fitting.tr_vector[3]
 
         dataset = MockObject()
         dataset._id = 97
-        r = ihm.restraint.CrossLinkRestraint(dataset=dataset, linker_type='DSS')
+        dss = ihm.ChemDescriptor('DSS')
+        r = ihm.restraint.CrossLinkRestraint(dataset=dataset, linker=dss)
         # intra, unambiguous
         xxl1 = ihm.restraint.ExperimentalCrossLink(e1.residue(2), e1.residue(3))
         # inter, ambiguous
@@ -2051,6 +2052,7 @@ _ihm_2dem_class_average_fitting.tr_vector[3]
 
         ihm.dumper._EntityDumper().finalize(system) # assign entity IDs
         ihm.dumper._StructAsymDumper().finalize(system) # assign asym IDs
+        ihm.dumper._ChemDescriptorDumper().finalize(system) # descriptor IDs
         dumper = ihm.dumper._CrossLinkDumper()
         dumper.finalize(system) # assign IDs
 
@@ -2067,12 +2069,13 @@ _ihm_cross_link_list.entity_description_2
 _ihm_cross_link_list.entity_id_2
 _ihm_cross_link_list.seq_id_2
 _ihm_cross_link_list.comp_id_2
+_ihm_cross_link_list.linker_descriptor_id
 _ihm_cross_link_list.linker_type
 _ihm_cross_link_list.dataset_list_id
-1 1 foo 1 2 THR foo 1 3 CYS DSS 97
-2 2 foo 1 2 THR bar 2 3 PHE DSS 97
-3 2 foo 1 2 THR bar 2 2 GLU DSS 97
-4 3 foo 1 1 ALA bar 2 1 ASP DSS 97
+1 1 foo 1 2 THR foo 1 3 CYS 1 DSS 97
+2 2 foo 1 2 THR bar 2 3 PHE 1 DSS 97
+3 2 foo 1 2 THR bar 2 2 GLU 1 DSS 97
+4 3 foo 1 1 ALA bar 2 1 ASP 1 DSS 97
 #
 #
 loop_