+++ test/test_reader.py
@@ -1708,14 +1708,14 @@ _ihm_cross_link_list.entity_description_2
 _ihm_cross_link_list.entity_id_2
 _ihm_cross_link_list.seq_id_2
 _ihm_cross_link_list.comp_id_2
-_ihm_cross_link_list.linker_type
+_ihm_cross_link_list.linker_descriptor_id
 _ihm_cross_link_list.dataset_list_id
-1 1 foo 1 2 THR foo 1 3 CYS DSS 97
-2 2 foo 1 2 THR bar 2 3 PHE DSS 97
-3 2 foo 1 2 THR bar 2 2 GLU DSS 97
-4 3 foo 1 1 ALA bar 2 1 ASP DSS 97
-5 4 foo 1 1 ALA bar 2 1 ASP EDC 97
-6 5 foo 1 1 ALA bar 2 1 ASP DSS 98
+1 1 foo 1 2 THR foo 1 3 CYS 44 97
+2 2 foo 1 2 THR bar 2 3 PHE 44 97
+3 2 foo 1 2 THR bar 2 2 GLU 44 97
+4 3 foo 1 1 ALA bar 2 1 ASP 44 97
+5 4 foo 1 1 ALA bar 2 1 ASP 88 97
+6 5 foo 1 1 ALA bar 2 1 ASP 44 98
 """)
         s, = ihm.reader.read(fh)
         # Check grouping
@@ -1723,7 +1723,7 @@ _ihm_cross_link_list.dataset_list_id
                           for r in s.restraints], [[1, 2, 1], [1], [1]])
         r1, r2, r3 = s.restraints
         self.assertEqual(r1.dataset._id, '97')
-        self.assertEqual(r1.linker_type, 'DSS')
+        self.assertEqual(r1.linker._id, '44')
         xl = r1.experimental_cross_links[1][0]
         self.assertEqual(xl.residue1.entity._id, '1')
         self.assertEqual(xl.residue2.entity._id, '2')