+++ ChangeLog.rst
@@ -1,7 +1,7 @@
 HEAD
 ====
  - Pseudo sites are now supported. :class:`ihm.restraint.PseudoSiteFeature`
-   always points or spheres with arbitrary coordinates to be designated as
+   allows points or spheres with arbitrary coordinates to be designated as
    features, which can then be used in
    :class:`ihm.restraint.DerivedDistanceRestraint`.